+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # CHANGELOG
 
+## [0.0.2] - 2019-04-18
+### Added
+- Service interfaces and application constants
+
 ## [0.0.1] - 2019-04-17
 ### Added
-- uuid package to abstract implementation
+- `uuid` package to abstract implementation
 
 ## [0.0.0] - 2019-04-12
 ### Added